Point symlinks to the files under version control (except .j2)

+++ _scripts/render-single.py
@@ -4,7 +4,6 @@ import argparse
 import logging
 import os
 from posixpath import dirname
-from shutil import copyfile
 import sys
 
 from jinja2 import Environment, FileSystemLoader
@@ -12,8 +11,8 @@ from jinja2 import Environment, FileSystemLoader
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
-    parser.add_argument("in_file")
-    parser.add_argument("out_file")
+    parser.add_argument("in_file", type=os.path.realpath)
+    parser.add_argument("out_file", type=os.path.realpath)
     args = parser.parse_args()
 
     logging.basicConfig(
@@ -24,10 +23,15 @@ if __name__ == "__main__":
 
     _, ext = os.path.splitext(args.in_file)
     if ext != ".j2":
-        logging.info(f"{args.in_file}: not a .j2 file, copying 1:1...")
         if os.path.exists(args.out_file):
-            os.remove(args.out_file)
-        copyfile(args.in_file, args.out_file)
+            if args.out_file == args.in_file:
+                logging.debug(
+                    f"{args.in_file}: in_file and out_file are the same...")
+                sys.exit(0)
+            else:
+                os.remove(args.out_file)
+        logging.debug(f"{args.in_file}: creating symlink...")
+        os.symlink(args.in_file, args.out_file)
         sys.exit(0)
 
     logging.info(f"{args.in_file}: rendering...")
